requirements.txt
tcod>=13.6.2
numpy>=1.22.4
Pillow>=9.1.1
snecs >= 1.2.2
